Moved type traits code into its own header file
