+++ library/include/rocwmma/internal/types_ext.hpp
@@ -99,319 +99,4 @@ namespace rocwmma
 
 } // namespace rocwmma
 
-namespace std
-{
-#if !defined(__HIPCC_RTC__)
-    ///////////////////////////////////////////////////////////
-    //////////  std::ostream::operator<<(float16_t)  //////////
-    ///////////////////////////////////////////////////////////
-
-    inline ostream& operator<<(ostream& stream, rocwmma::float16_t const& val)
-    {
-        return stream << static_cast<float>(val);
-    }
-
-    ///////////////////////////////////////////////////////////
-    //////////  std::ostream::operator<<(hfloat16_t)  /////////
-    ///////////////////////////////////////////////////////////
-
-    inline ostream& operator<<(ostream& stream, rocwmma::hfloat16_t const& val)
-    {
-        return stream << __half2float(val);
-    }
-#endif // !defined(__HIPCC_RTC__)
-
-#if defined(__HIPCC_RTC__)
-    using uint16_t = rocwmma::uint16_t;
-#endif
-
-    ///////////////////////////////////////////////////////////
-    ///////////  std::numeric_limits<float16_t>  //////////////
-    ///////////////////////////////////////////////////////////
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::epsilon() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x1400));
-        return eps.f16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::infinity() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7C00));
-        return eps.f16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::lowest() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0xFBFF));
-        return eps.f16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::max() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7BFF));
-        return eps.f16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::min() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x0400));
-        return eps.f16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::quiet_NaN() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7FFF));
-        return eps.f16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::float16_t
-        numeric_limits<rocwmma::float16_t>::signaling_NaN() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7DFF));
-        return eps.f16;
-    }
-
-    ///////////////////////////////////////////////////////////
-    ///////////  std::numeric_limits<hfloat16_t>  /////////////
-    ///////////////////////////////////////////////////////////
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::epsilon() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x1400));
-        return eps.h16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::infinity() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7C00));
-        return eps.h16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::lowest() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0xFBFF));
-        return eps.h16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::max() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7BFF));
-        return eps.h16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::min() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x0400));
-        return eps.h16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::quiet_NaN() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7FFF));
-        return eps.h16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::hfloat16_t
-        numeric_limits<rocwmma::hfloat16_t>::signaling_NaN() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7DFF));
-        return eps.h16;
-    }
-
-    ///////////////////////////////////////////////////////////
-    ///////////  std::numeric_limits<bfloat16_t>  /////////////
-    ///////////////////////////////////////////////////////////
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::epsilon() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x3C00));
-        return eps.b16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::infinity() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7F80));
-        return eps.b16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::lowest() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0xFF7F));
-        return eps.b16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::max() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7F7F));
-        return eps.b16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::min() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x007F));
-        return eps.b16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::quiet_NaN() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7FC0));
-        return eps.b16;
-    }
-
-    template <>
-    ROCWMMA_HOST_DEVICE constexpr rocwmma::bfloat16_t
-        numeric_limits<rocwmma::bfloat16_t>::signaling_NaN() noexcept
-    {
-        rocwmma::detail::Fp16Bits eps(static_cast<uint16_t>(0x7FC0));
-        return eps.b16;
-    }
-
-} // namespace std
-
-namespace rocwmma
-{
-#if !defined(__HIPCC_RTC__)
-    ///////////////////////////////////////////////////////////
-    ///////////  rocwmma::hfloat16_t host operators  //////////
-    ///////////////////////////////////////////////////////////
-
-    __host__ inline bool operator==(const hfloat16_t& x, const hfloat16_t& y)
-    {
-        auto absDiff = std::fabs(__half2float(x) - __half2float(y));
-        auto absAdd  = std::fabs(__half2float(x) + __half2float(y));
-        return absDiff <= __half2float(std::numeric_limits<hfloat16_t>::epsilon()) * absAdd * 2.0f
-               || absDiff < __half2float(std::numeric_limits<hfloat16_t>::min());
-    }
-
-    __host__ inline bool operator!=(const hfloat16_t& x, const hfloat16_t& y)
-    {
-        return !(x == y);
-    }
-
-    __host__ inline hfloat16_t operator-(const hfloat16_t& x)
-    {
-        detail::Fp16Bits fp16(x);
-        fp16.i16 ^= 0x8000; // Flip sign
-        return fp16.h16;
-    }
-
-    __host__ inline hfloat16_t operator+(const hfloat16_t& x, const hfloat16_t& y)
-    {
-        return static_cast<hfloat16_t>(static_cast<float16_t>(x) + static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t operator-(const hfloat16_t& x, const hfloat16_t& y)
-    {
-        return static_cast<hfloat16_t>(static_cast<float16_t>(x) - static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t operator*(const hfloat16_t& x, const hfloat16_t& y)
-    {
-        return static_cast<hfloat16_t>(static_cast<float16_t>(x) * static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t operator/(const hfloat16_t& x, const hfloat16_t& y)
-    {
-        return static_cast<hfloat16_t>(static_cast<float16_t>(x) / static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t& operator+=(hfloat16_t& x, const hfloat16_t& y)
-    {
-        return x = static_cast<hfloat16_t>(static_cast<float16_t>(x) + static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t& operator-=(hfloat16_t& x, const hfloat16_t& y)
-    {
-        return x = static_cast<hfloat16_t>(static_cast<float16_t>(x) - static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t& operator*=(hfloat16_t& x, const hfloat16_t& y)
-    {
-        return x = static_cast<hfloat16_t>(static_cast<float16_t>(x) * static_cast<float16_t>(y));
-    }
-
-    __host__ inline hfloat16_t& operator/=(hfloat16_t& x, const hfloat16_t& y)
-    {
-        return x = static_cast<hfloat16_t>(static_cast<float16_t>(x) / static_cast<float16_t>(y));
-    }
-
-    template <typename T, typename std::enable_if_t<std::is_integral<T>::value, int> = 0>
-    constexpr auto maxExactInteger() -> decltype(std::numeric_limits<T>::max())
-    {
-        return std::numeric_limits<T>::max();
-    }
-
-    template <typename T,
-              typename std::enable_if_t<std::is_floating_point<T>::value
-                                            && std::numeric_limits<T>::digits,
-                                        int>
-              = 0>
-    constexpr auto maxExactInteger() ->
-        typename std::conditional_t<std::is_same<T, float64_t>::value, int64_t, int32_t>
-    {
-        using RetT =
-            typename std::conditional_t<std::is_same<T, float64_t>::value, int64_t, int32_t>;
-        return ((RetT)1 << std::numeric_limits<T>::digits);
-    }
-
-    template <typename T,
-              typename std::enable_if_t<std::is_same<T, hfloat16_t>::value
-                                            || std::is_same<T, float16_t>::value,
-                                        int>
-              = 0>
-    constexpr auto maxExactInteger() -> int32_t
-    {
-        // f16 mantissa is 10 bits
-        return ((int32_t)1 << 11);
-    }
-
-    template <typename T, typename std::enable_if_t<std::is_same<T, bfloat16_t>::value, int> = 0>
-    constexpr auto maxExactInteger() -> int32_t
-    {
-        // b16 mantissa is 7 bits
-        return ((int32_t)1 << 8);
-    }
-#endif // !defined(__HIPCC_RTC__)
-
-} // namespace rocwmma
-
 #endif // ROCWMMA_TYPES_EXT_HPP
