+++ library/include/rocwmma/internal/utils.hpp
@@ -187,347 +187,8 @@ namespace std
 } // namespace std
 #endif // !defined(__HIPCC_RTC__)
 
-///////////////////////////////////////////////////////////
-/////////////  std replacements for hipRTC  ///////////////
-///////////////////////////////////////////////////////////
-#if defined(__HIPCC_RTC__)
-namespace std
-{
-    template <typename T>
-    class numeric_limits
-    {
-    public:
-        ROCWMMA_HOST_DEVICE static constexpr T min() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T lowest() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T max() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T epsilon() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T round_error() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T infinity() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T quiet_NaN() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T signaling_NaN() noexcept;
-        ROCWMMA_HOST_DEVICE static constexpr T denorm_min() noexcept;
-    };
-
-    template <bool B, class T = void>
-    using enable_if_t = typename enable_if<B, T>::type;
-
-    template <bool B, class T, class F>
-    struct conditional
-    {
-    };
-
-    template <class T, class F>
-    struct conditional<true, T, F>
-    {
-        using type = T;
-    };
-
-    template <class T, class F>
-    struct conditional<false, T, F>
-    {
-        using type = F;
-    };
-
-    template <bool B, class T, class F>
-    using conditional_t = typename conditional<B, T, F>::type;
-
-    template <typename T>
-    ROCWMMA_HOST_DEVICE const T& max(const T& a, const T& b)
-    {
-        return (a < b) ? b : a;
-    }
-
-    template <typename T>
-    ROCWMMA_HOST_DEVICE const T& min(const T& a, const T& b)
-    {
-        return (b < a) ? a : b;
-    }
-
-    // Meta programming helper types.
-
-    template <bool, typename, typename>
-    struct conditional;
-
-    template <typename...>
-    struct __or_;
-
-    template <>
-    struct __or_<> : public false_type
-    {
-    };
-
-    template <typename _B1>
-    struct __or_<_B1> : public _B1
-    {
-    };
-
-    template <typename _B1, typename _B2>
-    struct __or_<_B1, _B2> : public conditional<_B1::value, _B1, _B2>::type
-    {
-    };
-
-    template <typename _B1, typename _B2, typename _B3, typename... _Bn>
-    struct __or_<_B1, _B2, _B3, _Bn...>
-        : public conditional<_B1::value, _B1, __or_<_B2, _B3, _Bn...>>::type
-    {
-    };
-
-    template <typename...>
-    struct __and_;
-
-    template <>
-    struct __and_<> : public true_type
-    {
-    };
-
-    template <typename _B1>
-    struct __and_<_B1> : public _B1
-    {
-    };
-
-    template <typename _B1, typename _B2>
-    struct __and_<_B1, _B2> : public conditional<_B1::value, _B2, _B1>::type
-    {
-    };
-
-    template <typename _B1, typename _B2, typename _B3, typename... _Bn>
-    struct __and_<_B1, _B2, _B3, _Bn...>
-        : public conditional<_B1::value, __and_<_B2, _B3, _Bn...>, _B1>::type
-    {
-    };
-
-    template <bool __v>
-    using __bool_constant = integral_constant<bool, __v>;
-
-    template <typename _Pp>
-    struct __not_ : public __bool_constant<!bool(_Pp::value)>
-    {
-    };
-
-    // remove_reference
-    template <typename T>
-    struct remove_reference
-    {
-        typedef T type;
-    };
-
-    template <typename T>
-    struct remove_reference<T&>
-    {
-        typedef T type;
-    };
-
-    template <typename T>
-    struct remove_reference<T&&>
-    {
-        typedef T type;
-    };
-
-    // is_lvalue_reference
-    template <typename>
-    struct is_lvalue_reference : public false_type
-    {
-    };
-
-    template <typename T>
-    struct is_lvalue_reference<T&> : public true_type
-    {
-    };
-
-    // is_rvalue_reference
-    template <typename>
-    struct is_rvalue_reference : public false_type
-    {
-    };
-
-    template <typename T>
-    struct is_rvalue_reference<T&&> : public true_type
-    {
-    };
-
-    // lvalue forwarding
-    template <typename T>
-    constexpr T&& forward(typename remove_reference<T>::type& __t) noexcept
-    {
-        return static_cast<T&&>(__t);
-    }
-
-    // rvalue forwarding
-    template <typename T>
-    constexpr T&& forward(typename remove_reference<T>::type&& __t) noexcept
-    {
-        static_assert(!is_lvalue_reference<T>::value,
-                      "template argument"
-                      " substituting T is an lvalue reference type");
-        return static_cast<T&&>(__t);
-    }
-
-    // remove_const
-    template <typename T>
-    struct remove_const
-    {
-        typedef T type;
-    };
-
-    template <typename T>
-    struct remove_const<T const>
-    {
-        typedef T type;
-    };
-
-    // remove_volatile
-    template <typename T>
-    struct remove_volatile
-    {
-        typedef T type;
-    };
-
-    template <typename T>
-    struct remove_volatile<T volatile>
-    {
-        typedef T type;
-    };
-
-    // remove_cv
-    template <typename T>
-    struct remove_cv
-    {
-        typedef typename remove_const<typename remove_volatile<T>::type>::type type;
-    };
-
-    // remove_extent
-    template <typename T>
-    struct remove_extent
-    {
-        typedef T type;
-    };
-
-    template <typename T, std::size_t _Size>
-    struct remove_extent<T[_Size]>
-    {
-        typedef T type;
-    };
-
-    template <typename T>
-    struct remove_extent<T[]>
-    {
-        typedef T type;
-    };
-
-    // is_void
-    template <typename>
-    struct __is_void_helper : public false_type
-    {
-    };
-
-    template <>
-    struct __is_void_helper<void> : public true_type
-    {
-    };
-
-    template <typename T>
-    struct is_void : public __is_void_helper<typename remove_cv<T>::type>::type
-    {
-    };
-
-    // is_reference
-    template <typename T>
-    struct is_reference : public __or_<is_lvalue_reference<T>, is_rvalue_reference<T>>::type
-    {
-    };
-
-    // is_function
-    template <typename>
-    struct is_function : public false_type
-    {
-    };
-
-    // is_object
-    template <typename T>
-    struct is_object : public __not_<__or_<is_function<T>, is_reference<T>, is_void<T>>>::type
-    {
-    };
-
-    // __is_referenceable
-    template <typename T>
-    struct __is_referenceable : public __or_<is_object<T>, is_reference<T>>::type
-    {
-    };
-
-    // add_pointer
-    template <typename T, bool = __or_<__is_referenceable<T>, is_void<T>>::value>
-    struct __add_pointer_helper
-    {
-        typedef T type;
-    };
-
-    template <typename T>
-    struct __add_pointer_helper<T, true>
-    {
-        typedef typename remove_reference<T>::type* type;
-    };
-
-    template <typename T>
-    struct add_pointer : public __add_pointer_helper<T>
-    {
-    };
-
-    // is_array
-    template <typename>
-    struct is_array : public false_type
-    {
-    };
-
-    template <typename T, std::size_t _Size>
-    struct is_array<T[_Size]> : public true_type
-    {
-    };
-
-    template <typename T>
-    struct is_array<T[]> : public true_type
-    {
-    };
-
-    // decay selectors
-    template <typename _Up,
-              bool _IsArray    = is_array<_Up>::value,
-              bool _IsFunction = is_function<_Up>::value>
-    struct __decay_selector;
-
-    template <typename _Up>
-    struct __decay_selector<_Up, false, false>
-    {
-        typedef typename remove_cv<_Up>::type __type;
-    };
-
-    template <typename _Up>
-    struct __decay_selector<_Up, true, false>
-    {
-        typedef typename remove_extent<_Up>::type* __type;
-    };
-
-    template <typename _Up>
-    struct __decay_selector<_Up, false, true>
-    {
-        typedef typename add_pointer<_Up>::type __type;
-    };
-
-    // decay
-    template <typename T>
-    class decay
-    {
-        typedef typename remove_reference<T>::type __remove_type;
-
-    public:
-        typedef typename __decay_selector<__remove_type>::__type type;
-    };
-
-} // namespace std
-#endif
-
 namespace rocwmma
 {
-
     // Computes ceil(numerator/divisor) for integer types.
     template <typename intT1,
               class = typename std::enable_if<std::is_integral<intT1>::value>::type,
